Substantially reduce the thresholds at which the BTreePersistentComponents converts from dicts to trees.
This is intended to reduce pickle sizes and number of ghost objects.

The _subscribers attribute is actually special cased to always become a BTree, because for utilities with many registrations for the same interface, it will only have one level and the values in that level will be a single large tuple. This way we can at least defer loading it (many sites only have a few types of utilities registered). I think that for utilities _subscribers isn't actually used except when adding/removing utility registrations, so most read-only transactions shouldn't need to touch it.